+++ packages/kernel/src/learning/index.ts
@@ -2,5 +2,6 @@ export * from "./components";
 export * from "./pairing";
 export * from "./promotion";
 export * from "./sequential";
+export * from "./skill-evaluation";
 export * from "./types";
 export * from "./versions";